CUB/requirements.txt
GitPython==3.0.3
numpy==1.18.1
Pillow==7.1.2
pyaml==19.4.1
PyYAML==5.1.2
scipy==1.3.2
six==1.13.0
torch==1.5.0
torchvision==0.6.0
tqdm==4.38.0
